# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/semesters": {
            "get": {
                "description": "Get all semester labels (e.g. \"Spring 2024\")",
                "tags": [
                    "Semesters"
                ],
                "summary": "Get semester labels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create semester label (e.g. Spring 2024)",
                "tags": [
                    "Semesters"
                ],
                "summary": "Create semester label",
                "parameters": [
                    {
                        "description": "Semester Label (e.g. `Fall 2023`)",
                        "name": "semester",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/semesters/{semesterLabel}/statuses": {
            "get": {
                "description": "Get all brother statuses for a semester",
                "tags": [
                    "Semesters"
                ],
                "summary": "Get Brother statuses for a semester",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Attendance"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create all brother statuses for a semester",
                "tags": [
                    "Semesters"
                ],
                "summary": "Create Brother statuses for a semester",
                "parameters": [
                    {
                        "type": "string",
                        "description": "semesterLabel",
                        "name": "semesterLabel",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "BrotherID",
                        "name": "brotherID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Attendance"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.APIResponse": {
            "description": "JSON response format for all API calls",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Omit if nil"
                },
                "message": {
                    "description": "Omit if nil",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Attendance": {
            "type": "object",
            "properties": {
                "attendanceStatus": {
                    "type": "string"
                },
                "brotherID": {
                    "type": "integer"
                },
                "eventCategory": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "eventID": {
                    "type": "integer"
                },
                "eventLocation": {
                    "type": "string"
                },
                "eventName": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "rollCall": {
                    "type": "integer"
                }
            }
        },
...
    }
}